[Device detail fixes] This commit fixes Map in the device detail screen, timestamp  of graph and how values are exhibited in the graph.
