+++ src/js/components/HistoryElements.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import util from "../comms/util/util";
 import { Line } from 'react-chartjs-2';
+import { PositionRenderer } from '../views/devices/DeviceMap';
 
 
 class Graph extends Component{
@@ -33,7 +34,7 @@ class Graph extends Component{
 
     this.props.data[this.props.id][this.props.attr].map((i) => {
       labels.push(util.iso_to_date(i.ts));
-      values.push(i.trim());
+      values.push(i.value);
     })
 
     if (values.length === 0) {
@@ -137,6 +138,77 @@ function HistoryList(props) {
   return empty;
 }
 
+class PositionWrapper extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      opened: false,
+      hasPosition: false,
+      pos: []
+    };
+    this.getDevicesWithPosition = this.getDevicesWithPosition.bind(this);
+    this.toogleExpand = this.toogleExpand.bind(this);
+  }
+
+  toogleExpand(state) {
+    console.log("state", state);
+    this.setState({opened: state});
+  }
+
+
+  getDevicesWithPosition(device){
+    function parserPosition(position){
+      let parsedPosition = position.split(", ");
+      return [parseFloat(parsedPosition[0]), parseFloat(parsedPosition[1])];
+    }
+
+    let validDevices = [];
+    let length = device[this.props.attr].length;
+       for(let j in device.attrs){
+         for(let i in device.attrs[j]){
+           if(device.attrs[j][i].type == "static"){
+             if(device.attrs[j][i].value_type == "geo:point"){
+               device.position = parserPosition(device.attrs[j][i].static_value);
+             }
+           } else{
+             device.position = parserPosition(device[this.props.attr][0].value);
+           }
+         }
+       }
+
+      device.select = true;
+      if(device.position !== null && device.position !== undefined){
+        validDevices.push(device);
+      }
+    return validDevices;
+  }
+
+  render() {
+    function NoData() {
+        return (
+          <div className="valign-wrapper full-height background-info">
+            <div className="full-width center">No position <br />available</div>
+          </div>
+        )
+    }
+
+    console.log("Position Renderer ", this.props.device);
+    if (this.props.device === undefined)
+    {
+      return (<NoData />);
+    }
+
+    let validDevices = this.getDevicesWithPosition(this.props.data[this.props.id]);
+    console.log("validDevices", validDevices);
+    if (validDevices.length == 0) {
+      return <NoData />;
+    } else {
+      return <div className={"PositionRendererDiv " + (this.state.opened ? "expanded" : "compressed")}>
+          <PositionRenderer devices={validDevices} allowContextMenu={false} center={validDevices[0].position} />
+        </div>
+    }
+  }
+}
 
 function Attr(props) {
   const known = {
@@ -144,7 +216,8 @@ function Attr(props) {
     'float': Graph,
     'string': HistoryList,
     'geo': HistoryList,
-    'default': HistoryList
+    'default': HistoryList,
+    'geo:point': PositionWrapper
   };
 
   const Renderer = props.type in known ? known[props.type] : known['default'];
